testing/web-platform/tests/atomics/window-and-any-worker/wait/was-woken/was-woken.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<body>
  <iframe id="myiframe" style="display: none" src="../empty.html" ></iframe>
</body>

<script>

window.addEventListener("load", function(e) {
  async_test(function(t) {
    let worker = new Worker('wait-on-zero.js');
    worker.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, 'ok');
    });
    let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    worker.postMessage([ia.buffer]);
    setTimeout(() => assert_equals(1, Atomics.wake(ia, 0)), 500);
  }, "atomics-dedicated-worker-was-woken");
  async_test(function(t) {
    let worker = new Worker('was-woken-worker.js');
    worker.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, 'ok');
    });
    worker.postMessage([]);
  }, "atomics-dedicated-worker-was-woken-from-worker");
  async_test(function(t) {
    let worker = new SharedWorker('was-woken-shared-worker.js');
    worker.port.start();
    worker.port.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, 'ok');
    });
    worker.port.postMessage([]);
  }, "atomics-dedicated-worker-was-woken-from-shared-worker");
  async_test(function(t) {
    let iframe = document.getElementById('myiframe');
    let contentWindow = iframe.contentWindow;
    let test = this;
    contentWindow.onmessage = function(e) {
      let worker = new Worker('wait-on-zero.js');
      worker.onmessage = test.step_func_done(function(e) {
        assert_equals(e.data, 'ok');
      });
      let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
      worker.postMessage([ia.buffer]);
      setTimeout(() => assert_equals(1, Atomics.wake(ia, 0)), 500);
    }
    contentWindow.postMessage([], '*');
  }, "atomics-dedicated-worker-was-woken-from-iframe");
});

</script>
